Extract upgraderoom, topic, roomname, roomavatar to room-settings
